+++ supertokens_python/recipe/emailpassword/api/signin.py
@@ -19,11 +19,13 @@ if TYPE_CHECKING:
     from supertokens_python.recipe.emailpassword.interfaces import APIOptions, APIInterface
 
 from supertokens_python.exceptions import raise_bad_input_exception
+from supertokens_python.logger import LoggerCodes, debug_logger, info_logger
 
 from .utils import validate_form_fields_or_throw_error
 
 
 async def handle_sign_in_api(api_implementation: APIInterface, api_options: APIOptions):
+    info_logger("Calling SignInAPI")
     if api_implementation.disable_sign_in_post:
         return None
     body = await api_options.request.json()
@@ -34,6 +36,7 @@ async def handle_sign_in_api(api_implementation: APIInterface, api_options: APIO
                                                             form_fields_raw)
     response = await api_implementation.sign_in_post(form_fields, api_options, {})
 
+    debug_logger(response.status, LoggerCodes.ApiResponse)
     api_options.response.set_json_content(response.to_json())
 
     return api_options.response